use zcash msm when number of points less than 256
